Make the starting seed user configurable.
One case where this is useful is reproducible runs with impure tests, e.g.
when using proper to drive integration tests. Take the current time as seed,
print it out, and then pass {seed, Timestamp} to proper:module. If there is
a failure, you can replicate the whole run, even if individual tests are not
(completely) isolated, assuming they are still deterministic.